Create Navbar component.
