+++ src/App.js
@@ -1,11 +1,26 @@
 import './App.css';
+import {
+  createBrowserRouter,
+  createRoutesFromElements,
+  RouterProvider,
+  Route,
+} from 'react-router-dom';
+
+// Pages
+import RootLayout from './Components/RootLayout/RootLayout';
+import BooksPage from './pages/BooksPage/BooksPage';
+
+const router = createBrowserRouter(
+  createRoutesFromElements(
+    <Route path="/" element={<RootLayout />}>
+      <Route index path="books" element={<BooksPage />} />
+      {/* <Route path="categories" element={<CategoriesPage />} /> */}
+    </Route>,
+  ),
+);
 
 function App() {
-  return (
-    <div className="App">
-      <h1>Book Store</h1>
-    </div>
-  );
+  return <RouterProvider router={router} />;
 }
 
 export default App;
