+++ src/pages/BooksPage/BooksPage.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const BooksPage = () => (
+  <div>BooksPages</div>
+);
+
+export default BooksPage;
